refactor: separate queries
