+++ agent_api_rest/src/credential_issuer/well_known/openid_credential_issuer.rs
@@ -1,12 +1,14 @@
-use agent_issuance::{handlers::query_handler, server_config::queries::ServerConfigView, state::ApplicationState};
+use agent_issuance::{
+    handlers::query_handler,
+    server_config::queries::ServerConfigView,
+    state::{ApplicationState, SERVER_CONFIG_ID},
+};
 use axum::{
     extract::{Json, State},
     http::StatusCode,
     response::IntoResponse,
 };
 use tracing::info;
-
-use crate::SERVER_CONFIG_ID;
 
 #[axum_macros::debug_handler]
 pub(crate) async fn openid_credential_issuer(State(state): State<ApplicationState>) -> impl IntoResponse {
