+++ agent_api_rest/src/lib.rs
@@ -18,8 +18,6 @@ use credential_issuer::{
 use credentials::credentials;
 use offers::offers;
 
-pub const SERVER_CONFIG_ID: &str = "SERVER-CONFIG-001";
-
 pub fn app(app_state: ApplicationState) -> Router {
     let base_path = get_base_path();
