+++ agent_issuance/src/state.rs
@@ -9,13 +9,16 @@ use crate::credential::aggregate::Credential;
 use crate::credential::queries::CredentialView;
 use crate::handlers::command_handler;
 use crate::offer::aggregate::Offer;
-use crate::offer::queries::{AccessTokenView, OfferView, PreAuthorizedCodeView};
+use crate::offer::queries::access_token::AccessTokenView;
+use crate::offer::queries::pre_authorized_code::PreAuthorizedCodeView;
+use crate::offer::queries::OfferView;
 use crate::server_config::aggregate::ServerConfig;
 use crate::server_config::command::ServerConfigCommand;
 use crate::server_config::queries::ServerConfigView;
 
+/// The `Command` trait is used to define the command handlers for the aggregates.
 #[async_trait]
-pub trait CQRS<A>
+pub trait Command<A>
 where
     A: Aggregate,
 {
@@ -29,12 +32,14 @@ where
         A::Command: Send + Sync;
 }
 
+/// The application state is used to store the command handlers and queries.
 #[derive(Clone)]
 pub struct ApplicationState {
     pub command: CommandHandlers,
     pub query: Queries,
 }
 
+/// The command handlers are used to execute commands on the aggregates.
 #[derive(Clone)]
 pub struct CommandHandlers {
     pub server_config: CommandHandler<ServerConfig>,
@@ -42,8 +47,11 @@ pub struct CommandHandlers {
     pub offer: CommandHandler<Offer>,
 }
 
-pub type CommandHandler<A> = Arc<dyn CQRS<A> + Send + Sync>;
+pub type CommandHandler<A> = Arc<dyn Command<A> + Send + Sync>;
 
+/// This type is used to define the queries that are used to query the view repositories. We make use of `dyn` here, so
+/// that any type of repository that implements the `ViewRepository` trait can be used, but the corresponding `View` and
+/// `Aggregate` types must be the same.
 type Queries = ViewRepositories<
     dyn ViewRepository<ServerConfigView, ServerConfig>,
     dyn ViewRepository<CredentialView, Credential>,
@@ -79,6 +87,7 @@ impl Clone for Queries {
     }
 }
 
+/// Returns a new `GenericQuery` instance.
 pub fn generic_query<R, A, V>(view_repository: Arc<R>) -> GenericQuery<R, V, A>
 where
     R: ViewRepository<V, A>,
@@ -91,13 +100,16 @@ where
     generic_query
 }
 
+/// The unique identifier for the server configuration.
+pub const SERVER_CONFIG_ID: &str = "SERVER-CONFIG-001";
+
 /// Initialize the application state by executing the startup commands.
 pub async fn initialize(state: ApplicationState, startup_commands: Vec<ServerConfigCommand>) {
     info!("Initializing ...");
 
     for command in startup_commands {
         let command_string = format!("{:?}", command).split(' ').next().unwrap().to_string();
-        match command_handler("SERVER-CONFIG-001", &state.command.server_config, command).await {
+        match command_handler(SERVER_CONFIG_ID, &state.command.server_config, command).await {
             Ok(_) => info!("Startup task completed: `{}`", command_string),
             Err(err) => warn!("Startup task failed: {:#?}", err),
         }
