+++ agent_store/src/postgres.rs
@@ -1,12 +1,11 @@
 use agent_issuance::{
     credential::services::CredentialServices,
     offer::{
-        aggregate::Offer,
-        queries::{AccessTokenView, OfferSubQuery, PreAuthorizedCodeView},
+        queries::{access_token::AccessTokenQuery, pre_authorized_code::PreAuthorizedCodeQuery},
         services::OfferServices,
     },
     server_config::services::ServerConfigServices,
-    state::{generic_query, ApplicationState, CommandHandlers, ViewRepositories, CQRS},
+    state::{generic_query, ApplicationState, Command, CommandHandlers, ViewRepositories},
 };
 use agent_shared::config;
 use async_trait::async_trait;
@@ -23,7 +22,7 @@ where
 }
 
 #[async_trait]
-impl<A> CQRS<A> for AggregateHandler<A>
+impl<A> Command<A> for AggregateHandler<A>
 where
     A: Aggregate + 'static,
     <A as Aggregate>::Command: Send + Sync,
@@ -68,13 +67,9 @@ pub async fn application_state() -> agent_issuance::state::ApplicationState {
     let pre_authorized_code = Arc::new(PostgresViewRepository::new("pre_authorized_code", pool.clone()));
     let access_token = Arc::new(PostgresViewRepository::new("access_token", pool.clone()));
 
-    // Create sub-queries for the offer aggregate.
-    let pre_authorized_code_query: OfferSubQuery<
-        PostgresViewRepository<PreAuthorizedCodeView, Offer>,
-        PreAuthorizedCodeView,
-    > = OfferSubQuery::new(pre_authorized_code.clone(), "pre-authorized_code".to_string());
-    let access_token_query: OfferSubQuery<PostgresViewRepository<AccessTokenView, Offer>, AccessTokenView> =
-        OfferSubQuery::new(access_token.clone(), "access_token".to_string());
+    // Create custom-queries for the offer aggregate.
+    let pre_authorized_code_query = PreAuthorizedCodeQuery::new(pre_authorized_code.clone());
+    let access_token_query = AccessTokenQuery::new(access_token.clone());
 
     ApplicationState {
         command: CommandHandlers {
